re-organized tests to make it easier to skip when working on a single one
