+++ tests/e2e/e2e_bank_test.go
@@ -1,0 +1,59 @@
+package e2e
+
+import (
+	"fmt"
+	"time"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+func (s *IntegrationTestSuite) testBankTokenTransfer() {
+	s.Run("send_photon_between_accounts", func() {
+		var err error
+		senderAddress := s.chainA.validators[0].keyInfo.GetAddress()
+		sender := senderAddress.String()
+
+		recipientAddress := s.chainA.validators[1].keyInfo.GetAddress()
+		recipient := recipientAddress.String()
+
+		chainAAPIEndpoint := fmt.Sprintf("http://%s", s.valResources[s.chainA.id][0].GetHostPort("1317/tcp"))
+
+		var (
+			beforeSenderUAtomBalance    sdk.Coin
+			beforeRecipientUAtomBalance sdk.Coin
+		)
+
+		s.Require().Eventually(
+			func() bool {
+				beforeSenderUAtomBalance, err = getSpecificBalance(chainAAPIEndpoint, sender, uatomDenom)
+				s.Require().NoError(err)
+
+				beforeRecipientUAtomBalance, err = getSpecificBalance(chainAAPIEndpoint, recipient, uatomDenom)
+				s.Require().NoError(err)
+
+				return beforeSenderUAtomBalance.IsValid() && beforeRecipientUAtomBalance.IsValid()
+			},
+			10*time.Second,
+			5*time.Second,
+		)
+
+		s.execBankSend(s.chainA, 0, sender, recipient, tokenAmount.String(), standardFees.String(), false)
+
+		s.Require().Eventually(
+			func() bool {
+				afterSenderUAtomBalance, err := getSpecificBalance(chainAAPIEndpoint, sender, uatomDenom)
+				s.Require().NoError(err)
+
+				afterRecipientUAtomBalance, err := getSpecificBalance(chainAAPIEndpoint, recipient, uatomDenom)
+				s.Require().NoError(err)
+
+				decremented := beforeSenderUAtomBalance.Sub(tokenAmount).Sub(standardFees).IsEqual(afterSenderUAtomBalance)
+				incremented := beforeRecipientUAtomBalance.Add(tokenAmount).IsEqual(afterRecipientUAtomBalance)
+
+				return decremented && incremented
+			},
+			time.Minute,
+			5*time.Second,
+		)
+	})
+}
