+++ tests/e2e/e2e_bypassminfee_test.go
@@ -1,0 +1,22 @@
+package e2e
+
+import (
+	"cosmossdk.io/math"
+)
+
+func (s *IntegrationTestSuite) testByPassMinFeeWithdrawReward() {
+	paidFeeAmt := math.LegacyMustNewDecFromStr(minGasPrice).Mul(math.LegacyNewDec(gas)).String()
+	payee := s.chainA.validators[0].keyInfo.GetAddress()
+	// pass
+	s.T().Logf("bypass-msg with fee in the denom of global fee, pass")
+	s.execWithdrawAllRewards(s.chainA, 0, payee.String(), paidFeeAmt+uatomDenom, false)
+	// pass
+	s.T().Logf("bypass-msg with zero coin in the denom of global fee, pass")
+	s.execWithdrawAllRewards(s.chainA, 0, payee.String(), "0"+uatomDenom, false)
+	// pass
+	s.T().Logf("bypass-msg with zero coin not in the denom of global fee, pass")
+	s.execWithdrawAllRewards(s.chainA, 0, payee.String(), "0"+photonDenom, false)
+	// fail
+	s.T().Logf("bypass-msg with non-zero coin not in the denom of global fee, fail")
+	s.execWithdrawAllRewards(s.chainA, 0, payee.String(), paidFeeAmt+photonDenom, true)
+}
