+++ tests/e2e/e2e_staking_test.go
@@ -1,17 +1,32 @@
 package e2e
 
 import (
+	"fmt"
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-func (s *IntegrationTestSuite) testStaking(chainEndpoint string, delegatorAddress string, validatorAddressA string, validatorAddressB string, fees sdk.Coin, homePath string) {
+func (s *IntegrationTestSuite) testStaking() {
+	chainEndpoint := fmt.Sprintf("http://%s", s.valResources[s.chainA.id][0].GetHostPort("1317/tcp"))
+
+	validatorA := s.chainA.validators[0]
+	validatorB := s.chainA.validators[1]
+	validatorAAddr := validatorA.keyInfo.GetAddress()
+	validatorBAddr := validatorB.keyInfo.GetAddress()
+
+	validatorAddressA := sdk.ValAddress(validatorAAddr).String()
+	validatorAddressB := sdk.ValAddress(validatorBAddr).String()
+
+	delegatorAddress := s.chainA.genesisAccounts[2].keyInfo.GetAddress().String()
+
+	fees := sdk.NewCoin(uatomDenom, sdk.NewInt(10))
+
 	delegationAmount := sdk.NewInt(500000000)
 	delegation := sdk.NewCoin(uatomDenom, delegationAmount) // 500 atom
 
 	// Alice delegate uatom to Validator A
-	s.executeDelegate(s.chainA, 0, delegation.String(), validatorAddressA, delegatorAddress, homePath, fees.String())
+	s.executeDelegate(s.chainA, 0, delegation.String(), validatorAddressA, delegatorAddress, gaiaHomePath, fees.String())
 
 	// Validate delegation successful
 	s.Require().Eventually(
@@ -27,7 +42,7 @@ func (s *IntegrationTestSuite) testStaking(chainEndpoint string, delegatorAddres
 	)
 
 	// Alice re-delegate uatom from Validator A to Validator B
-	s.executeRedelegate(s.chainA, 0, delegation.String(), validatorAddressA, validatorAddressB, delegatorAddress, homePath, fees.String())
+	s.executeRedelegate(s.chainA, 0, delegation.String(), validatorAddressA, validatorAddressB, delegatorAddress, gaiaHomePath, fees.String())
 
 	// Validate re-delegation successful
 	s.Require().Eventually(
